+++ ApptentiveConnect/source/Message Center/ATFileMessageUserCellV7.m
@@ -20,6 +20,7 @@
 - (void)setup {
 	self.userIconView.image = [ATBackend imageNamed:@"at_mc_user_icon"];
 	self.userIconView.imageURL = [NSURL URLWithString:self.message.sender.profilePhotoURL];
+	
 	self.userIconView.layer.cornerRadius = self.userIconView.bounds.size.width*0.5;
 	self.userIconView.layer.masksToBounds = YES;