+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Fixes:
 
 * [Issue #65](https://github.com/apptentive/apptentive-ios/issues/65) Long response is truncated in message centre on the device
 * Fixes a crash caused by setting link attributes after text on TTTAttributedLabels.
+* Fixes IOS-332 Message Center avatars are fetched from server on every load
 
 2013-11-22 wooster, pkamb v1.2.3
 --------------------------------